+++ example/vite/src/App.tsx
@@ -6,6 +6,7 @@ import {
   Select as S,
   Text,
   Button,
+  Link,
 } from "@kuma-ui/core";
 import { Dynamic } from "./Dynamic";
 import { MouseEventHandler, useEffect, useRef } from "react";
@@ -30,7 +31,18 @@ function App() {
     <Box color="GrayText" border="1px solid" borderColor="orange">
       test
       <Styled>Styled</Styled>
+      <Text contain="size">日本語</Text>
       <Dynamic />
+      <Link
+        href="https://www.kuma-ui.com/"
+        target="_blank"
+        _after={{
+          content: '"🚀"',
+          fontSize: 24,
+        }}
+      >
+        Go to Our Website
+      </Link>
     </Box>
   );
 }