+++ packages/babel-plugin/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @kuma-ui/babel-plugin
 
+## 1.2.3
+
+### Patch Changes
+
+- Updated dependencies [[`fa14ed8`](https://github.com/kuma-ui/kuma-ui/commit/fa14ed8039cd966d8cb7841c21bac45a6d5e16ed), [`ba2e216`](https://github.com/kuma-ui/kuma-ui/commit/ba2e216832663f7a303ff1e4445389b1122825a1)]:
+  - @kuma-ui/system@1.7.6
+  - @kuma-ui/core@1.5.9
+
+## 1.2.2
+
+### Patch Changes
+
+- Updated dependencies [[`67c9006`](https://github.com/kuma-ui/kuma-ui/commit/67c9006b1b196e94c6e05eab45ac315caf557742), [`fdba9cc`](https://github.com/kuma-ui/kuma-ui/commit/fdba9cc7c760a6652d2bb345867adfb5288946a3), [`5b17f23`](https://github.com/kuma-ui/kuma-ui/commit/5b17f2397499f5f9e0eeacc4632201ac59b10f3b), [`6055686`](https://github.com/kuma-ui/kuma-ui/commit/60556869e5c3ceecd8fa2ef9c52dd90dd4be4ad1)]:
+  - @kuma-ui/core@1.5.8
+  - @kuma-ui/system@1.7.5
+
 ## 1.2.1
 
 ### Patch Changes