+++ packages/vite/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @kuma-ui/vite
 
+## 1.4.0
+
+### Minor Changes
+
+- [#427](https://github.com/kuma-ui/kuma-ui/pull/427) [`42ab3c3`](https://github.com/kuma-ui/kuma-ui/commit/42ab3c31106ba6d3db3e4e3061d4a3902e68f510) Thanks [@yossydev](https://github.com/yossydev)! - vite v6 support
+
+### Patch Changes
+
+- Updated dependencies [[`fa14ed8`](https://github.com/kuma-ui/kuma-ui/commit/fa14ed8039cd966d8cb7841c21bac45a6d5e16ed)]:
+  - @kuma-ui/system@1.7.6
+  - @kuma-ui/compiler@1.3.3
+
+## 1.3.2
+
+### Patch Changes
+
+- Updated dependencies [[`fdba9cc`](https://github.com/kuma-ui/kuma-ui/commit/fdba9cc7c760a6652d2bb345867adfb5288946a3), [`2b1b7ce`](https://github.com/kuma-ui/kuma-ui/commit/2b1b7ce98c64a9fc7b59c99f381f8046cd306b3a), [`5b17f23`](https://github.com/kuma-ui/kuma-ui/commit/5b17f2397499f5f9e0eeacc4632201ac59b10f3b)]:
+  - @kuma-ui/system@1.7.5
+  - @kuma-ui/compiler@1.3.2
+
 ## 1.3.1
 
 ### Patch Changes