+++ packages/wasm/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @kuma-ui/wasm
 
+## 1.0.3
+
+### Patch Changes
+
+- [#420](https://github.com/kuma-ui/kuma-ui/pull/420) [`2b1b7ce`](https://github.com/kuma-ui/kuma-ui/commit/2b1b7ce98c64a9fc7b59c99f381f8046cd306b3a) Thanks [@poteboy](https://github.com/poteboy)! - Update oxc From v0.12.5 to v0.25.0
+
 ## 1.0.2
 
 ### Patch Changes