+++ packages/webpack-plugin/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @kuma-ui/webpack-plugin
 
+## 1.4.3
+
+### Patch Changes
+
+- Updated dependencies [[`fa14ed8`](https://github.com/kuma-ui/kuma-ui/commit/fa14ed8039cd966d8cb7841c21bac45a6d5e16ed)]:
+  - @kuma-ui/system@1.7.6
+  - @kuma-ui/compiler@1.3.3
+
+## 1.4.2
+
+### Patch Changes
+
+- Updated dependencies [[`fdba9cc`](https://github.com/kuma-ui/kuma-ui/commit/fdba9cc7c760a6652d2bb345867adfb5288946a3), [`2b1b7ce`](https://github.com/kuma-ui/kuma-ui/commit/2b1b7ce98c64a9fc7b59c99f381f8046cd306b3a), [`5b17f23`](https://github.com/kuma-ui/kuma-ui/commit/5b17f2397499f5f9e0eeacc4632201ac59b10f3b)]:
+  - @kuma-ui/system@1.7.5
+  - @kuma-ui/compiler@1.3.2
+
 ## 1.4.1
 
 ### Patch Changes